BanditLab/pyproject.toml
[tool.poetry]
name = "mab-framework"
version = "0.1.0"
description = ""
authors = ["e.suraveykin <surikkatik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
pyyaml = "^6.0.3"
numpy = "^2.4.4"
matplotlib = "^3.10.8"
pandas = "^3.0.2"
scikit-learn = "^1.8.0"
coba = "^8.0.5"
torch = "^2.11.0"
cvxpy = "^1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
